Refactoring for better UI logic isolation ahead of further logic enhancements
